% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd.R
\name{xkcd}
\alias{xkcd}
\title{metadata about an xkcd comic}
\usage{
xkcd(number)
}
\description{
Given an xkcd comic number, this function grabs a json object that is associated with that number. it displays the comic in the R studio plot window
}
